Stopped material handling for Edwards2019 drag
Implements a static friction constraint that balances forces locally to
hold arrested material in place if the net force on the flow is below
the threshold for motion.

Adds a new user flag, Stopped Material Handling = on/off.

Limitations: only 1D for now and only for this particular drag law. Not
tested with morphodynamics, since this isn't the intended use case.
Furthermore, the implementation requires that the friction is treated
explicitly, so we revert to the (second order) Heun's method in this
case.